+++ tests/cli_commands/test_init.py
@@ -12,7 +12,7 @@ from data_pipelines_cli.cli_constants import DEFAULT_GLOBAL_CONFIG
 from data_pipelines_cli.data_structures import (
     DataPipelinesConfig,
     TemplateConfig,
-    read_config,
+    read_env_config,
 )
 from data_pipelines_cli.errors import DataPipelinesError
 
@@ -40,10 +40,10 @@ class InitCommandTestCase(unittest.TestCase):
     def test_init(self):
         runner = CliRunner()
         with tempfile.TemporaryDirectory() as tmp_dir, patch(
-            "data_pipelines_cli.cli_commands.init.CONFIGURATION_PATH",
+            "data_pipelines_cli.cli_commands.init.ENV_CONFIGURATION_PATH",
             pathlib.Path(tmp_dir).joinpath(".dp.yml"),
         ), patch(
-            "data_pipelines_cli.cli_constants.CONFIGURATION_PATH",
+            "data_pipelines_cli.cli_constants.ENV_CONFIGURATION_PATH",
             pathlib.Path(tmp_dir).joinpath(".dp.yml"),
         ):
             result = runner.invoke(
@@ -52,20 +52,20 @@ class InitCommandTestCase(unittest.TestCase):
                 input="test_user\n/var/tmp",
             )
             self.assertEqual(0, result.exit_code, msg=result.exception)
-            self.assertEqual(self.example_config_dict, read_config())
+            self.assertEqual(self.example_config_dict, read_env_config())
 
     def test_global_config(self):
         runner = CliRunner()
         with tempfile.TemporaryDirectory() as tmp_dir, patch(
-            "data_pipelines_cli.cli_commands.init.CONFIGURATION_PATH",
+            "data_pipelines_cli.cli_commands.init.ENV_CONFIGURATION_PATH",
             pathlib.Path(tmp_dir).joinpath(".dp.yml"),
         ), patch(
-            "data_pipelines_cli.cli_constants.CONFIGURATION_PATH",
+            "data_pipelines_cli.cli_constants.ENV_CONFIGURATION_PATH",
             pathlib.Path(tmp_dir).joinpath(".dp.yml"),
         ):
             result = runner.invoke(_cli, ["init"])
             self.assertEqual(0, result.exit_code, msg=result.exception)
-            self.assertEqual(DEFAULT_GLOBAL_CONFIG, read_config())
+            self.assertEqual(DEFAULT_GLOBAL_CONFIG, read_env_config())
 
     @patch("questionary.confirm")
     def test_overwrite_yes(self, mock_questionary):
@@ -76,17 +76,17 @@ class InitCommandTestCase(unittest.TestCase):
         with tempfile.TemporaryDirectory() as tmp_dir:
             tmp_dp_path = pathlib.Path(tmp_dir).joinpath(".dp.yml")
             with patch(
-                "data_pipelines_cli.cli_commands.init.CONFIGURATION_PATH",
+                "data_pipelines_cli.cli_commands.init.ENV_CONFIGURATION_PATH",
                 tmp_dp_path,
             ), patch(
-                "data_pipelines_cli.cli_constants.CONFIGURATION_PATH",
+                "data_pipelines_cli.cli_constants.ENV_CONFIGURATION_PATH",
                 tmp_dp_path,
             ):
                 with open(tmp_dp_path, "w") as tmp_file:
                     yaml.dump(self.example_config_dict, tmp_file)
-                    self.assertEqual(self.example_config_dict, read_config())
+                    self.assertEqual(self.example_config_dict, read_env_config())
                 init(None)
-                self.assertEqual(DEFAULT_GLOBAL_CONFIG, read_config())
+                self.assertEqual(DEFAULT_GLOBAL_CONFIG, read_env_config())
 
     @patch("questionary.confirm")
     def test_overwrite_no(self, mock_questionary):
@@ -97,16 +97,16 @@ class InitCommandTestCase(unittest.TestCase):
         with tempfile.TemporaryDirectory() as tmp_dir:
             tmp_dp_path = pathlib.Path(tmp_dir).joinpath(".dp.yml")
             with patch(
-                "data_pipelines_cli.cli_commands.init.CONFIGURATION_PATH",
+                "data_pipelines_cli.cli_commands.init.ENV_CONFIGURATION_PATH",
                 tmp_dp_path,
             ), patch(
-                "data_pipelines_cli.cli_constants.CONFIGURATION_PATH",
+                "data_pipelines_cli.cli_constants.ENV_CONFIGURATION_PATH",
                 tmp_dp_path,
             ):
                 with open(tmp_dp_path, "w") as tmp_file:
                     yaml.dump(self.example_config_dict, tmp_file)
-                    self.assertEqual(self.example_config_dict, read_config())
+                    self.assertEqual(self.example_config_dict, read_env_config())
 
                 with self.assertRaises(DataPipelinesError):
                     init(None)
-                self.assertEqual(self.example_config_dict, read_config())
+                self.assertEqual(self.example_config_dict, read_env_config())